requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-ratelimit
django-redis
celery
redis
opencv-python
dlib
face_recognition
face_recognition_models
Pillow
numpy
reportlab
django-crispy-forms
crispy-bootstrap5
django-htmx
sentry-sdk
channels
channels-redis
geopy
twilio
qrcode[pil]
tailwindcss
